test: split workflow auto-advance e2e coverage
