+++ apps/web/e2e/tests/workflow/workflow-automation.spec.ts
@@ -1,179 +1,7 @@
 import { test, expect } from "../../fixtures/test-base";
-import { KanbanPage } from "../../pages/kanban-page";
 import { SessionPage } from "../../pages/session-page";
 
 test.describe("Workflow automation", () => {
-  /**
-   * Seeds a 3-step workflow (Inbox → In Progress → Done).
-   * Configures In Progress with on_enter: auto_start_agent and a custom step
-   * prompt routed to the mock agent, plus on_turn_complete: move_to_step(Done).
-   *
-   * Verifies:
-   * - auto_start_agent fires when task is moved into In Progress
-   * - task appears in In Progress column before agent completes
-   * - step custom prompt is used (not the task description)
-   * - on_turn_complete advances task to Done column on the kanban board
-   * - session stepper shows Done as the current step
-   */
-  test("auto_start_agent with step custom prompt; on_turn_complete advances to next column", async ({
-    testPage,
-    apiClient,
-    seedData,
-  }) => {
-    // --- Seed a workflow with explicit step configuration ---
-    const workflow = await apiClient.createWorkflow(
-      seedData.workspaceId,
-      "Automation Test Workflow",
-    );
-
-    const inboxStep = await apiClient.createWorkflowStep(workflow.id, "Inbox", 0);
-    const inProgressStep = await apiClient.createWorkflowStep(workflow.id, "In Progress", 1);
-    const doneStep = await apiClient.createWorkflowStep(workflow.id, "Done", 2);
-
-    // Configure In Progress: auto_start_agent on enter, custom step prompt with delay
-    // so we can observe the task in In Progress before on_turn_complete moves it to Done.
-    await apiClient.updateWorkflowStep(inProgressStep.id, {
-      prompt: 'e2e:delay(3000)\ne2e:message("delayed step response")\n{{task_prompt}}',
-      events: {
-        on_enter: [{ type: "auto_start_agent" }],
-        on_turn_complete: [{ type: "move_to_step", config: { step_id: doneStep.id } }],
-      },
-    });
-
-    // Point the kanban page at our custom workflow
-    await apiClient.saveUserSettings({
-      workspace_id: seedData.workspaceId,
-      workflow_filter_id: workflow.id,
-      enable_preview_on_click: false,
-    });
-
-    // Use the seed profile so tests don't pick up passthrough profiles from other tests.
-    const agentProfileId = seedData.agentProfileId;
-
-    // Create task in Inbox — does NOT trigger auto_start_agent
-    const task = await apiClient.createTask(seedData.workspaceId, "Auto Agent Workflow Task", {
-      workflow_id: workflow.id,
-      workflow_step_id: inboxStep.id,
-      agent_profile_id: agentProfileId,
-      repository_ids: [seedData.repositoryId],
-    });
-
-    // Move task to In Progress → triggers on_enter: auto_start_agent
-    await apiClient.moveTask(task.id, workflow.id, inProgressStep.id);
-
-    const kanban = new KanbanPage(testPage);
-    await kanban.goto();
-
-    // Task should appear in In Progress while the agent is working (3s delay)
-    const cardInProgress = kanban.taskCardInColumn("Auto Agent Workflow Task", inProgressStep.id);
-    await expect(cardInProgress).toBeVisible({ timeout: 15_000 });
-
-    // After the agent completes its turn, on_turn_complete fires and moves the
-    // task to Done. The kanban board receives a WS push.
-    const cardInDone = kanban.taskCardInColumn("Auto Agent Workflow Task", doneStep.id);
-    await expect(cardInDone).toBeVisible({ timeout: 30_000 });
-
-    // Navigate to the session page
-    await cardInDone.click();
-    await expect(testPage).toHaveURL(/\/t\//, { timeout: 15_000 });
-
-    const session = new SessionPage(testPage);
-    await session.waitForLoad();
-
-    // Stepper shows Done as current step
-    await expect(session.stepperStep("Done")).toHaveAttribute("aria-current", "step", {
-      timeout: 10_000,
-    });
-
-    // Mock agent response confirms the step custom prompt was executed
-    await expect(session.chat.getByText("delayed step response", { exact: true })).toBeVisible({
-      timeout: 30_000,
-    });
-
-    // Session transitions to idle after agent completes
-    await session.waitForChatIdle({ timeout: 15_000 });
-
-    // Sidebar shows the task under "Turn Finished" section
-    await expect(session.sidebarSection("Turn Finished")).toBeVisible();
-  });
-
-  test("does not retrigger previous step prompt after auto-advancing into Done", async ({
-    testPage,
-    apiClient,
-    seedData,
-  }) => {
-    test.setTimeout(120_000);
-
-    const workflow = await apiClient.createWorkflow(
-      seedData.workspaceId,
-      "Final Step No Retrigger Workflow",
-    );
-
-    const startStep = await apiClient.createWorkflowStep(workflow.id, "Start", 0);
-    const workStep = await apiClient.createWorkflowStep(workflow.id, "Work", 1);
-    await apiClient.createWorkflowStep(workflow.id, "Done", 2);
-
-    await apiClient.updateWorkflowStep(startStep.id, {
-      events: {
-        on_turn_complete: [{ type: "move_to_next" }],
-      },
-    });
-
-    const workMarker = "final-step-work-response-once";
-    await apiClient.updateWorkflowStep(workStep.id, {
-      prompt: `e2e:message("${workMarker}")\n{{task_prompt}}`,
-      events: {
-        on_enter: [{ type: "auto_start_agent" }],
-        on_turn_complete: [{ type: "move_to_next" }],
-      },
-    });
-
-    await apiClient.saveUserSettings({
-      workspace_id: seedData.workspaceId,
-      workflow_filter_id: workflow.id,
-      enable_preview_on_click: false,
-    });
-
-    const task = await apiClient.createTask(seedData.workspaceId, "Final Done Retrigger Task", {
-      workflow_id: workflow.id,
-      workflow_step_id: startStep.id,
-      agent_profile_id: seedData.agentProfileId,
-      repository_ids: [seedData.repositoryId],
-    });
-
-    const session = new SessionPage(testPage);
-    await testPage.goto(`/t/${task.id}`);
-    await session.waitForLoad();
-    await expect(session.stepperStep("Start")).toHaveAttribute("aria-current", "step", {
-      timeout: 10_000,
-    });
-
-    await session.sendMessageViaButton("/e2e:simple-message");
-
-    await expect(session.stepperStep("Done")).toHaveAttribute("aria-current", "step", {
-      timeout: 45_000,
-    });
-    await expect(session.chat.getByText(workMarker, { exact: true })).toBeVisible({
-      timeout: 30_000,
-    });
-    await session.waitForChatIdle({ timeout: 30_000 });
-
-    const taskSessions = await apiClient.listTaskSessions(task.id);
-    const primarySession = taskSessions.sessions.find((candidate) => candidate.is_primary);
-    expect(primarySession, "expected a primary task session").toBeTruthy();
-
-    const { messages } = await apiClient.listSessionMessages(primarySession!.id);
-    const workResponses = messages.filter(
-      (message) => message.author_type === "agent" && message.content.includes(workMarker),
-    );
-    const workPrompts = messages.filter(
-      (message) => message.author_type === "user" && message.content.includes(workMarker),
-    );
-
-    expect(workResponses).toHaveLength(1);
-    expect(workPrompts).toHaveLength(1);
-  });
-
   /**
    * Seeds a 4-step workflow (Todo → In Progress → Review → Done) and validates
    * the full lifecycle from both the kanban page and the session details page.
